# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CyberSafe Teens API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Lesson {
  id: number;
  slug: string;
  level: number;
  titleEn: string;
  titleUz: string;
  titleRu: string;
  descriptionEn: string;
  descriptionUz: string;
  descriptionRu: string;
  xpReward: number;
  estimatedMinutes: number;
  /** @nullable */
  icon?: string | null;
}

export interface LessonDetail {
  id: number;
  slug: string;
  level: number;
  titleEn: string;
  titleUz: string;
  titleRu: string;
  contentEn: string;
  contentUz: string;
  contentRu: string;
  realExampleEn: string;
  realExampleUz: string;
  realExampleRu: string;
  actionStepsEn: string;
  actionStepsUz: string;
  actionStepsRu: string;
  xpReward: number;
  estimatedMinutes: number;
  /** @nullable */
  icon?: string | null;
}

export type UserProgressQuizScores = {[key: string]: number};

export interface UserProgress {
  username: string;
  completedLessons: number[];
  totalXp: number;
  currentStreak: number;
  badges: string[];
  quizScores: UserProgressQuizScores;
}

export interface LessonCompleteInput {
  username: string;
  lessonId: number;
  /** @nullable */
  quizScore?: number | null;
}

export interface QuizOption {
  en: string;
  uz: string;
  ru: string;
}

export interface QuizQuestion {
  id: number;
  lessonId: number;
  questionEn: string;
  questionUz: string;
  questionRu: string;
  options: QuizOption[];
  correctIndex: number;
  explanationEn: string;
  explanationUz: string;
  explanationRu: string;
}

export interface QuizSubmission {
  username: string;
  lessonId: number;
  answers: number[];
}

export interface QuizFeedback {
  questionId: number;
  correct: boolean;
  explanationEn: string;
  explanationUz: string;
  explanationRu: string;
}

export interface QuizResult {
  score: number;
  total: number;
  passed: boolean;
  xpEarned: number;
  feedback: QuizFeedback[];
}

export interface SimulatorScenario {
  id: number;
  type: string;
  platform: string;
  messageEn: string;
  messageUz: string;
  messageRu: string;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  senderAvatar?: string | null;
  isScam: boolean;
  warningSignsEn: string;
  warningSignsUz: string;
  warningSignsRu: string;
  explanationEn: string;
  explanationUz: string;
  explanationRu: string;
}

export interface SimulatorAnswerInput {
  scenarioId: number;
  username: string;
  answeredScam: boolean;
}

export interface SimulatorAnswerResult {
  correct: boolean;
  isScam: boolean;
  explanationEn: string;
  explanationUz: string;
  explanationRu: string;
  warningSignsEn: string;
  warningSignsUz: string;
  warningSignsRu: string;
  xpEarned: number;
}

export interface PasswordCheckInput {
  password: string;
}

export interface PasswordCheckResult {
  score: number;
  label: string;
  timeToCrackEn: string;
  timeToCrackUz: string;
  timeToCrackRu: string;
  tipsEn: string[];
  tipsUz: string[];
  tipsRu: string[];
  hasUppercase: boolean;
  hasLowercase: boolean;
  hasNumbers: boolean;
  hasSymbols: boolean;
  isCommon: boolean;
  length: number;
}

export interface ScamReport {
  id: number;
  message: string;
  category: string;
  platform: string;
  reportedAt: string;
  upvotes: number;
  /** @nullable */
  warningEn?: string | null;
  /** @nullable */
  warningUz?: string | null;
  /** @nullable */
  warningRu?: string | null;
}

export interface ScamReportInput {
  message: string;
  category: string;
  platform: string;
  /** @nullable */
  reporterName?: string | null;
}

export interface ScamStat {
  category: string;
  count: number;
}

export interface LeaderboardEntry {
  rank: number;
  username: string;
  totalXp: number;
  completedLessons: number;
  badges: string[];
  currentStreak: number;
}

export interface PlatformStats {
  totalUsers: number;
  totalLessonsCompleted: number;
  totalScamsReported: number;
  totalQuizzesTaken: number;
}

export type GetUserProgressParams = {
username: string;
};

export type GetScamReportsParams = {
category?: string;
limit?: number;
};

export type GetLeaderboardParams = {
limit?: number;
};