# lib/api-zod/src/generated/types/quizOption.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CyberSafe Teens API
 * OpenAPI spec version: 0.1.0
 */

export interface QuizOption {
  en: string;
  uz: string;
  ru: string;
}